refactor(`ios`): split _ios.py into _base.py, _pgsql_osm.py, and _pbf_importer.py
